+++ NEWS.md
@@ -1,7 +1,12 @@
-# Version 1.1 (2016-02-18)
+# Version 1.1.1 (2016-02-19)
+
+  - Fixed function for latex compilation by checking latex flavor and type of OS
+
+# Version 1.1.0 (2016-02-18)
 
   - Redefined pdflatex command (excluding sintex)
   - Fixed bugs in vignette and examples
   - Improved text of vignette
-
-# Version 1.0 - Initial version
+  - Fixed function for latex compilation by checking latex flavor and type of OS
+  
+# Version 1.0.0 - Initial version (2016-02-01)